Time: 0 ms (100.00%) | Memory: 7.3 MB (98.39%) - LeetSync

+++ 119-pascals-triangle-ii/pascals-triangle-ii.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     vector<int> getRow(int n) {
      vector<int> result(n+1,1);
-     for(int i=1;i<n;i++){
-         result[i]=(long long)result[i-1]*(n-i+1)/i;
+     for(long long i=1;i<n;i++){
+         result[i]=result[i-1]*(n-i+1)/i;
      }
      return result;
     }
